+++ antivoice/common.h
