+++ antivoice/mic_input.h
