+++ antivoice/mode_of_operation.h
