--- globals.h
@@ -1,13 +1,0 @@
-#ifndef COMMON_H
-#define COMMON_H
-
-/* This file contains all general includes and type declarations */
-#include "stdio.h"
-#include "string.h"
-
-typedef unsigned char bool_t;
-#define FALSE 0
-#define TRUE 1
-
-
-#endif /*COMMON_H*/
