--- main_algorithm.h
@@ -1,7 +1,0 @@
-#ifndef MAIN_ALGORITHM_H
-#define MAIN_ALGORITHM_H
-
-#include "common.h"
-
-
-#endif /* MAIN_ALGORITHM_H */
